Added functionality for vendor transactions. Some additional minor changes.

+++ Data/src/main/java/com/data/entities/CollegeProgramFeeMap.java
@@ -19,7 +19,7 @@ public class CollegeProgramFeeMap {
 	private Integer id;
 	
 	@Column(name="Amount")
-	private Integer amount;
+	private Float amount;
 	
 	@Column(name="OnlyForDisplayPurpose")
 	private Boolean onlyForDisplayPurpose;
@@ -47,11 +47,11 @@ public class CollegeProgramFeeMap {
 			this.id = feeId;
 		}
 	
-		public Integer getAmount() {
+		public Float getAmount() {
 			return amount;
 		}
 	
-		public void setAmount(Integer amount) {
+		public void setAmount(Float amount) {
 			this.amount = amount;
 		}
